+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Log Subscribe/StartTrial events going through GooglePlay store, which was logged as fb_mobile_purchase
+
 ## [4.41.0] - 2019-03-08
 
 ### Removed